licensing/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pyjwt==2.10.1
pydantic[email]==2.10.3
jinja2==3.1.4
python-multipart==0.0.20
httpx==0.28.1
psycopg2-binary==2.9.10
cryptography==44.0.0
